refactor: remove pkg/resources dependencies on wgtypes, netx
This finishes the work started in #4469, #4524.

The only thing left for #4420 is to move the package in the source tree.
