+++ internal/app/machined/pkg/adapters/kubespan/kubespan.go
@@ -2,5 +2,5 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// Package perf implements adapters wrapping resources/perf to provide additional functionality.
-package perf
+// Package kubespan implements adapters wrapping resources/kubespan to provide additional functionality.
+package kubespan
