+++ internal/app/machined/pkg/adapters/kubespan/peer_status_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"inet.af/netaddr"
 
+	kubespanadapter "github.com/talos-systems/talos/internal/app/machined/pkg/adapters/kubespan"
 	"github.com/talos-systems/talos/pkg/resources/kubespan"
 )
 
@@ -19,7 +20,7 @@ func TestPeerStatus_PickNewEndpoint(t *testing.T) {
 	peerStatus := kubespan.PeerStatusSpec{}
 
 	// no endpoint => no way to pick new one
-	assert.True(t, peerStatus.PickNewEndpoint(nil).IsZero())
+	assert.True(t, kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(nil).IsZero())
 
 	endpoints := []netaddr.IPPort{
 		netaddr.MustParseIPPort("10.3.4.5:10500"),
@@ -27,36 +28,36 @@ func TestPeerStatus_PickNewEndpoint(t *testing.T) {
 	}
 
 	// initial choice should be the first endpoint
-	newEndpoint := peerStatus.PickNewEndpoint(endpoints)
+	newEndpoint := kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints)
 	assert.Equal(t, endpoints[0], newEndpoint)
-	peerStatus.UpdateEndpoint(newEndpoint)
+	kubespanadapter.PeerStatusSpec(&peerStatus).UpdateEndpoint(newEndpoint)
 
 	// next choice should be 2nd endpoint
-	newEndpoint = peerStatus.PickNewEndpoint(endpoints)
+	newEndpoint = kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints)
 	assert.Equal(t, endpoints[1], newEndpoint)
-	peerStatus.UpdateEndpoint(newEndpoint)
+	kubespanadapter.PeerStatusSpec(&peerStatus).UpdateEndpoint(newEndpoint)
 
 	// back to the first endpoint
-	newEndpoint = peerStatus.PickNewEndpoint(endpoints)
+	newEndpoint = kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints)
 	assert.Equal(t, endpoints[0], newEndpoint)
-	peerStatus.UpdateEndpoint(newEndpoint)
+	kubespanadapter.PeerStatusSpec(&peerStatus).UpdateEndpoint(newEndpoint)
 
 	// can't rotate a single endpoint
-	assert.True(t, peerStatus.PickNewEndpoint(endpoints[:1]).IsZero())
+	assert.True(t, kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints[:1]).IsZero())
 
 	// can rotate if the endpoint is different
-	newEndpoint = peerStatus.PickNewEndpoint(endpoints[1:])
+	newEndpoint = kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints[1:])
 	assert.Equal(t, endpoints[1], newEndpoint)
-	peerStatus.UpdateEndpoint(newEndpoint)
+	kubespanadapter.PeerStatusSpec(&peerStatus).UpdateEndpoint(newEndpoint)
 
 	// if totally new list of endpoints is given, pick the first one
 	endpoints = []netaddr.IPPort{
 		netaddr.MustParseIPPort("10.3.4.5:10501"),
 		netaddr.MustParseIPPort("192.168.3.8:458"),
 	}
-	newEndpoint = peerStatus.PickNewEndpoint(endpoints)
+	newEndpoint = kubespanadapter.PeerStatusSpec(&peerStatus).PickNewEndpoint(endpoints)
 	assert.Equal(t, endpoints[0], newEndpoint)
-	peerStatus.UpdateEndpoint(newEndpoint)
+	kubespanadapter.PeerStatusSpec(&peerStatus).UpdateEndpoint(newEndpoint)
 }
 
 func TestPeerStatus_CalculateState(t *testing.T) {
@@ -78,38 +79,38 @@ func TestPeerStatus_CalculateState(t *testing.T) {
 		},
 		{
 			name:                "peer is down",
-			sinceLastHandshake:  2 * kubespan.PeerDownInterval,
-			sinceEndpointChange: 2 * kubespan.PeerDownInterval,
+			sinceLastHandshake:  2 * kubespanadapter.PeerDownInterval,
+			sinceEndpointChange: 2 * kubespanadapter.PeerDownInterval,
 			expectedState:       kubespan.PeerStateDown,
 		},
 		{
 			name:                "fresh peer, no handshake",
-			sinceLastHandshake:  2 * kubespan.PeerDownInterval,
-			sinceEndpointChange: kubespan.EndpointConnectionTimeout / 2,
+			sinceLastHandshake:  2 * kubespanadapter.PeerDownInterval,
+			sinceEndpointChange: kubespanadapter.EndpointConnectionTimeout / 2,
 			expectedState:       kubespan.PeerStateUnknown,
 		},
 		{
 			name:                "fresh peer, with handshake",
 			sinceLastHandshake:  0,
-			sinceEndpointChange: kubespan.EndpointConnectionTimeout / 2,
+			sinceEndpointChange: kubespanadapter.EndpointConnectionTimeout / 2,
 			expectedState:       kubespan.PeerStateUp,
 		},
 		{
 			name:                "peer after initial timeout, with handshake",
 			sinceLastHandshake:  0,
-			sinceEndpointChange: kubespan.EndpointConnectionTimeout + 1,
+			sinceEndpointChange: kubespanadapter.EndpointConnectionTimeout + 1,
 			expectedState:       kubespan.PeerStateUp,
 		},
 		{
 			name:                "peer after initial timeout, no handshake",
-			sinceLastHandshake:  2 * kubespan.EndpointConnectionTimeout,
-			sinceEndpointChange: kubespan.EndpointConnectionTimeout + 1,
+			sinceLastHandshake:  2 * kubespanadapter.EndpointConnectionTimeout,
+			sinceEndpointChange: kubespanadapter.EndpointConnectionTimeout + 1,
 			expectedState:       kubespan.PeerStateDown,
 		},
 		{
 			name:                "established peer, up",
-			sinceLastHandshake:  kubespan.PeerDownInterval / 2,
-			sinceEndpointChange: kubespan.PeerDownInterval + 1,
+			sinceLastHandshake:  kubespanadapter.PeerDownInterval / 2,
+			sinceEndpointChange: kubespanadapter.PeerDownInterval + 1,
 			expectedState:       kubespan.PeerStateUp,
 		},
 	} {
@@ -125,7 +126,7 @@ func TestPeerStatus_CalculateState(t *testing.T) {
 				peerStatus.LastUsedEndpoint = netaddr.MustParseIPPort("192.168.1.1:10000")
 			}
 
-			peerStatus.CalculateStateWithDurations(tt.sinceLastHandshake, tt.sinceEndpointChange)
+			kubespanadapter.PeerStatusSpec(&peerStatus).CalculateStateWithDurations(tt.sinceLastHandshake, tt.sinceEndpointChange)
 
 			assert.Equal(t, tt.expectedState, peerStatus.State)
 		})
